# Improve `Interval` constructor performance